fix nested udt_array

+++ v2/converters/other_types_test.go
@@ -14,7 +14,7 @@ Typ=100 Len=4: 60,249,140,204
 func TestBinaryFloat(t *testing.T) {
 	cases := []struct {
 		raw      []byte
-		expected float32
+		expected float64
 	}{
 		{
 			[]byte{195, 6, 115, 51},
